learned conditional rendering techniques
react ignores booleans and null so can use that and some laws of conditional return data to conditionally return info. learned ternary statements, function conditional returns, and inline boolean conditional returns: true && 'data' returns data. the second boolean is returned. if false, the first boolean is returned. if that first boolean IS the bool false then nothing is displayed. if the second bool IS html/data, then it displays the data